# Closes #1343 On-demand encode/decode of Operation values during chain resolution