The chain resolution now handles the following cases:

1. Decode value of target operations only when required
2. Avoid decoding values of non-target operations
3. No expiry calculations and expiryTime updates when expiry is not configured
4. Avoid rebuilding the chain after resolution if no compaction has happened