@lorban 2016-08-12
Maybe a few JHM tests would be handy to figure out how much extra performance this change gives us. And a Flight Recorder log would also help making sure this code is squeezing as much as it can from the hardware.