@albinsuresh 2016-08-17
The test failures are due to the platform version updates. I never updated this PR after the version changes were reverted. Thought of doing it after the JMH test additions are also done.