@albinsuresh 2016-08-18
The following branch has the JMH test added for the resolve operation: https://github.com/albinsuresh/ehcache3/tree/jmh

The average time for the _resolve_ operation was measured with the following configuration:
```
# JMH 1.12 (released 139 days ago, please consider updating!)
# VM version: JDK 1.8.0_51, VM 25.51-b03
# VM invoker: /Library/Java/JavaVirtualMachines/jdk1.8.0_51.jdk/Contents/Home/jre/bin/java
# VM options: -Dfile.encoding=UTF-8 -Duser.country=US -Duser.language=en -Duser.variant
# Warmup: 10 iterations, 1 s each
# Measurement: 10 iterations, 1 s each
# Timeout: 10 min per iteration
# Threads: 1 thread, will synchronize iterations
# Benchmark mode: Average time, time/op
# Benchmark: org.ehcache.clustered.client.internal.store.operations.ChainResolverJMHTest.measureResolve
```

And the results were as follows:

Before the optimization:
```
# Run progress: 0.00% complete, ETA 00:00:20
# Fork: 1 of 1
# Warmup Iteration   1: 2.349 us/op
# Warmup Iteration   2: 1.941 us/op
# Warmup Iteration   3: 1.600 us/op
# Warmup Iteration   4: 1.660 us/op
# Warmup Iteration   5: 1.671 us/op
# Warmup Iteration   6: 1.620 us/op
# Warmup Iteration   7: 2.027 us/op
# Warmup Iteration   8: 1.917 us/op
# Warmup Iteration   9: 1.766 us/op
# Warmup Iteration  10: 1.723 us/op
Iteration   1: 1.712 us/op
Iteration   2: 1.690 us/op
Iteration   3: 1.677 us/op
Iteration   4: 1.666 us/op
Iteration   5: 1.666 us/op
Iteration   6: 1.714 us/op
Iteration   7: 1.686 us/op
Iteration   8: 1.678 us/op
Iteration   9: 1.688 us/op
Iteration  10: 1.675 us/op


Result "measureResolve":
  1.685 ±(99.9%) 0.025 us/op [Average]
  (min, avg, max) = (1.666, 1.685, 1.714), stdev = 0.017
  CI (99.9%): [1.660, 1.711] (assumes normal distribution)


# Run complete. Total time: 00:00:20

Benchmark                            Mode  Cnt  Score   Error  Units
ChainResolverJMHTest.measureResolve  avgt   10  1.685 ± 0.025  us/op
```

After optimization:
```
# Run progress: 0.00% complete, ETA 00:00:20
# Fork: 1 of 1
# Warmup Iteration   1: 1.136 us/op
# Warmup Iteration   2: 1.194 us/op
# Warmup Iteration   3: 0.888 us/op
# Warmup Iteration   4: 0.852 us/op
# Warmup Iteration   5: 0.851 us/op
# Warmup Iteration   6: 0.827 us/op
# Warmup Iteration   7: 0.855 us/op
# Warmup Iteration   8: 0.870 us/op
# Warmup Iteration   9: 0.837 us/op
# Warmup Iteration  10: 0.864 us/op
Iteration   1: 0.878 us/op
Iteration   2: 0.874 us/op
Iteration   3: 0.847 us/op
Iteration   4: 0.844 us/op
Iteration   5: 0.859 us/op
Iteration   6: 0.827 us/op
Iteration   7: 0.829 us/op
Iteration   8: 0.824 us/op
Iteration   9: 0.840 us/op
Iteration  10: 0.830 us/op


Result "measureResolve":
  0.845 ±(99.9%) 0.030 us/op [Average]
  (min, avg, max) = (0.824, 0.845, 0.878), stdev = 0.020
  CI (99.9%): [0.816, 0.875] (assumes normal distribution)


# Run complete. Total time: 00:00:20

Benchmark                            Mode  Cnt  Score   Error  Units
ChainResolverJMHTest.measureResolve  avgt   10  0.845 ± 0.030  us/op
```