@ljacomet 2016-08-18
Good to merge once rebased